Improve separation of concerns for the client
