+++ src/Network/GRPC/Client/Call.hs
@@ -1,16 +1,15 @@
--- | Internal state of an open RPC call
+-- | Open (ongoing) RPC call
 --
 -- Intended for uqnqualified import.
 module Network.GRPC.Client.Call (
     -- * Definition
-    IsFinal(..)
-  , Trailers(..)
-  , Call(..)
-  , InboundQ
-  , OutboundQ
+    Call -- opaque
     -- * Construction
   , openCall
   , closeCall
+    -- * Open call
+  , sendInput
+  , recvOutput
     -- * Logging
   , LogRPC(..)
   ) where
@@ -221,6 +220,89 @@ data CallClosed = CallClosed {
       callClosedAt     :: CallStack
     , callClosedReason :: Maybe SomeException
     }
+  deriving stock (Show)
+  deriving anyclass (Exception)
+
+{-------------------------------------------------------------------------------
+  Open call
+-------------------------------------------------------------------------------}
+
+-- | Send an input to the peer
+--
+-- This lives in @STM@ for improved composability. For example, if the peer is
+-- currently busy then 'sendInput' will block, but you can then use 'orElse' to
+-- provide an alternative codepath.
+--
+-- Calling 'sendInput' again after sending the final message is a bug.
+sendInput ::
+     Call rpc
+  -> IsFinal
+  -> Maybe (Input rpc)
+  -> STM ()
+sendInput Call{callOutbound, callSentFinal} isFinal input = do
+    sentFinal <- readTVar callSentFinal
+
+    if sentFinal then
+      throwSTM SendAfterFinal
+    else do
+      q <- getThreadInterface callOutbound
+
+      -- We write to the queue and update 'callSentFinal' in the same STM tx;
+      -- this is important to avoid race conditions.
+      case isFinal of
+        Final    -> writeTVar callSentFinal True
+        NotFinal -> return ()
+
+      -- By checking that we haven't sent the final message yet, we know that
+      -- this call to 'putMVar' will not block indefinitely: the thread that
+      -- sends messages to the peer will get to it eventually (unless it dies,
+      -- in which case the thread status will change and the call to
+      -- 'getThreadInterface' will be retried).
+      putTMVar q (isFinal, input)
+
+-- | Receive an output from the peer
+--
+-- This lives in @STM@ for improved compositionality. For example, you can wait
+-- on multiple clients and see which one responds first.
+--
+-- Though the types cannot guarantee it, you will receive 'Trailers' once, after
+-- the final 'Output'. Calling 'recvOutput' again after receiving 'Trailers' is
+-- a bug.
+recvOutput :: Call rpc -> STM (Either Trailers (Output rpc))
+recvOutput Call{callInbound, callRecvTrailers} = do
+    recvTrailers <- readTVar callRecvTrailers
+
+    if recvTrailers then
+      throwSTM RecvAfterTrailers
+    else do
+      q <- getThreadInterface callInbound
+
+      -- By checking that we haven't received the trailers yet, we know that
+      -- this call to 'takeTMVar' will not block indefinitely: the thread that
+      -- receives messages from the peer will get to it eventually (unless it
+      -- dies, in which case the thread status will change and the call to
+      -- 'getThreadInterface' will be retried).
+      r <- takeTMVar q
+
+      -- We read from the queue and update 'calRecvTrailers' in the same STM tx;
+      -- this is important to avoid race conditions.
+      case r of
+        Left _trailers -> writeTVar callRecvTrailers True
+        Right _output  -> return ()
+
+      return r
+
+-- | Thrown by 'sendInput'
+--
+-- This indicates a bug in client code. See 'sendInput' for discussion.
+data SendAfterFinal = SendAfterFinal
+  deriving stock (Show)
+  deriving anyclass (Exception)
+
+-- | Thrown by 'recvOutput'
+--
+-- This indicates a bug in client code. See 'recvOutput' for discussion.
+data RecvAfterTrailers = RecvAfterTrailers
   deriving stock (Show)
   deriving anyclass (Exception)
